udev-utils.c: Ignore "legacy" devices when EVDEV is enabled
When EVDEV is enabled in the kernel, devices are exposed twice:
* once through EVDEV (`/dev/input/event*`)
* once through a device-specific driver (e.g. `/dev/atkbd*`)

Users probably want a single one to be exposed and detected. Otherwise,
they may need more rules in xorg.conf(5) to "blacklist" duplicates.

So if EVDEV is enabled, we assume in libudev-devd that they want to use
the EVDEV devices, not the "legacy" ones. This simplifies the
input devices configuration.

Note that this doesn't prevent the "conflict" between individual devices
and multiplexers (kbdmux(8) and sysmouse(8)).