Add support for sysctl-based evdev property enumeration
A big security advantage of Wayland is not allowing applications to
read input devices all the time. Having /dev/input/* accessible to
the user account subverts this advantage.

libudev-devd was opening the evdev devices to detect their types (mouse,
keyboard, touchpad, etc). This does not work when /dev/input/* is
inaccessible.

With the kernel exposing this information via sysctls, we can work without
/dev/input/* access, preserving the Wayland security model.